requirements.txt
h5py
pycryptodome
paho-mqtt
psutil